pyproject.toml
[project]
name = "facturai"
version = "1.0.0"
description = "Servicio de extracción de datos de facturas argentinas usando Claude API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pdf2image>=1.17.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["src*"]
